# acp-writer/src/acp_writer/api.py
"""FastAPI REST API for the acp-writer service."""

import base64
import json
import logging
import os
import xml.etree.ElementTree as ET
from datetime import datetime, timezone

import mlflow
import requests
from fastapi import FastAPI, HTTPException, Request, Response
from fastapi.responses import JSONResponse

from cpg_contracts import (
    CPGMetadata,
    DecisionModelSummary,
    DecisionVariable,
    Recommendation,
    RecommendationBundle,
    RecommendationSearchRequest,
)

from acp_writer.store.embedding import (
    EmbeddingProvider,
    make_embedding_provider,
)
from acp_writer.store.guidelines_store import GuidelinesStore
from acp_writer.store.vector_store import InMemoryVectorStore, VectorStore

try:
    mlflow.fastapi.autolog()
except AttributeError:
    pass

logger = logging.getLogger(__name__)

KOGITO_URL = os.environ.get("KOGITO_URL", "http://localhost:8081")

# DMN metadata is parsed with namespace-wildcard matches ("{*}tag") so it works
# across DMN language versions (1.3, 1.4, …) without pinning a MODEL namespace.

_dynamic_models: dict[str, dict] = {}

# --- Store initialization ---
# The provider is chosen by the EMBEDDING_PROVIDER env switch
# (make_embedding_provider): "openai" for a real OpenAI-compatible endpoint,
# otherwise a FakeEmbeddingProvider (the default — no network/downloads on
# import, so tests stay hermetic). Call init_stores() to swap providers.

_embedding_provider: EmbeddingProvider = make_embedding_provider()
_vector_store: VectorStore = InMemoryVectorStore(_embedding_provider)
_guidelines_store: GuidelinesStore = GuidelinesStore(_vector_store)


def init_stores(embedding_provider: EmbeddingProvider | None = None) -> None:
    """Re-initialize stores with a specific embedding provider."""
    global _embedding_provider, _vector_store, _guidelines_store
    if embedding_provider:
        _embedding_provider = embedding_provider
    _vector_store = InMemoryVectorStore(_embedding_provider)
    _guidelines_store = GuidelinesStore(_vector_store)


app = FastAPI(
    title="ACP Writer API",
    version="0.2.0",
    description="Composes patient-specific, FHIR-compliant care plans.",
)



@app.on_event("startup")
async def startup():
    logging.basicConfig(
        level=logging.INFO,
        format="%(asctime)s %(levelname)-5s %(name)s: %(message)s",
        datefmt="%H:%M:%S",
        force=True,
    )


def _check_kogito() -> bool:
    try:
        r = requests.get(f"{KOGITO_URL}/q/health/ready", timeout=5)
        return r.status_code == 200
    except requests.RequestException:
        return False


import re as _re

_CODE_TOKEN_RE = _re.compile(r"(https?://[^\s|]+)\|(\S+)")


def _extract_codes(input_data_el: ET.Element) -> list[str]:
    """Extract clinical codes from a DMN inputData element.

    Looks for codes in two places (in order):
    1. extensionElements with code annotations (future cpg-ingester output)
    2. Structured code tokens in the description element text

    Code format: "system-url|code" (e.g. "http://loinc.org|8480-6").
    """
    codes: list[str] = []

    ext = input_data_el.find("{*}extensionElements")
    if ext is not None:
        for child in ext:
            system = child.get("system", "")
            code = child.get("code", "")
            if system and code:
                codes.append(f"{system}|{code}")
            elif child.text:
                codes.extend(_CODE_TOKEN_RE.findall(child.text))

    if not codes:
        desc_el = input_data_el.find("{*}description")
        if desc_el is not None and desc_el.text:
            for system, code in _CODE_TOKEN_RE.findall(desc_el.text):
                codes.append(f"{system}|{code}")

    return codes


def _extract_description(input_data_el: ET.Element) -> str | None:
    """Extract description text from a DMN inputData element."""
    desc_el = input_data_el.find("{*}description")
    if desc_el is not None and desc_el.text:
        return desc_el.text.strip()
    return None


def _parse_dmn_metadata(dmn_xml: str) -> DecisionModelSummary:
    """Extract model name, inputs, and outputs from DMN XML."""
    root = ET.fromstring(dmn_xml)

    model_name = root.get("name", "unknown")
    model_id = model_name.lower().replace(" ", "-")

    inputs = []
    for input_data in root.findall("{*}inputData"):
        var = input_data.find("{*}variable")
        if var is not None:
            codes = _extract_codes(input_data)
            desc = _extract_description(input_data)
            inputs.append(DecisionVariable(
                name=var.get("name", ""),
                type=var.get("typeRef", "string"),
                codes=codes or None,
                description=desc,
            ))

    if not inputs:
        seen_names: set[str] = set()
        for decision in root.findall("{*}decision"):
            dt = decision.find("{*}decisionTable")
            if dt is not None:
                for inp in dt.findall("{*}input"):
                    input_expr = inp.find("{*}inputExpression")
                    if input_expr is not None:
                        text_el = input_expr.find("{*}text")
                        var_name = text_el.text.strip() if text_el is not None and text_el.text else inp.get("label", "")
                        if var_name and var_name not in seen_names:
                            seen_names.add(var_name)
                            inputs.append(DecisionVariable(
                                name=var_name,
                                type=input_expr.get("typeRef", "string"),
                            ))

    outputs = []
    for decision in root.findall("{*}decision"):
        dt = decision.find("{*}decisionTable")
        if dt is not None:
            for output in dt.findall("{*}output"):
                outputs.append(DecisionVariable(
                    name=output.get("name", ""),
                    type=output.get("typeRef", "string"),
                ))

    return DecisionModelSummary(
        id=model_id,
        name=model_name,
        inputs=inputs,
        outputs=outputs,
        deployed_at=datetime.now(timezone.utc),
    )


@mlflow.trace(span_type="TOOL", name="evaluate_jit_dmn")
def _evaluate_jit(dmn_xml: str, inputs: dict) -> dict:
    """Evaluate DMN via the JIT endpoint on the decision-service."""
    from acp_writer.tools.dmn_evaluation import DmnEngineError

    dmn_b64 = base64.b64encode(dmn_xml.encode()).decode()
    r = requests.post(
        f"{KOGITO_URL}/jit/dmn",
        json={"dmn_xml_base64": dmn_b64, "inputs": inputs},
        timeout=30,
    )
    if 400 <= r.status_code < 500:
        raise DmnEngineError.from_response(r)
    r.raise_for_status()
    return r.json()


@mlflow.trace(span_type="TOOL", name="validate_dmn_with_engine")
def _validate_dmn_with_engine(dmn_xml: str) -> dict | None:
    """Validate DMN with KIE, returning ``None`` when the engine is unavailable.

    Deployment remains available when the optional decision engine is down. The
    caller distinguishes that fail-open path from an engine response with
    ``valid: false`` and rejects only the latter.
    """
    dmn_b64 = base64.b64encode(dmn_xml.encode()).decode()
    try:
        response = requests.post(
            f"{KOGITO_URL}/jit/dmn/validate",
            json={"dmn_xml_base64": dmn_b64},
            timeout=30,
        )
        response.raise_for_status()
        result = response.json()
        if not isinstance(result, dict) or "valid" not in result:
            raise ValueError("decision engine returned an invalid validation response")
        return result
    except (requests.RequestException, ValueError) as exc:
        logger.warning("DMN engine validation unavailable; accepting model provisionally: %s", exc)
        return None


def _validation_failure_response(validation: dict) -> JSONResponse:
    messages = validation.get("messages", [])
    return JSONResponse(
        status_code=422,
        content={
            "error": "DMN engine validation failed",
            "messages": messages,
        },
    )


# --- Health ---


@app.get("/health")
def health():
    return {"status": "UP"}


@app.get("/health/ready")
def readiness():
    if _check_kogito():
        return {"status": "UP"}
    raise HTTPException(status_code=503, detail="Decision engine not available")


@app.get("/api/v1/status")
def status():
    kogito_healthy = _check_kogito()
    total_models = len(_dynamic_models)
    return {
        "version": "0.2.0",
        "decision_engine": {
            "status": "healthy" if kogito_healthy else "unavailable",
            "models_deployed": total_models,
        },
        "knowledge_base": {
            "status": "available",
            "guidelines_registered": _guidelines_store.count(),
            "recommendations_ingested": _vector_store.count(),
        },
    }


# --- Care Plans ---


@app.post("/api/v1/careplans", status_code=201)
async def generate_careplan(request: Request):
    bundle = await request.json()

    if bundle.get("resourceType") != "Bundle":
        raise HTTPException(status_code=400, detail="Request body must be a FHIR Bundle")

    import uuid

    from acp_writer.checkpointer import get_checkpointer
    from acp_writer.pipeline import build_pipeline

    litellm_url = os.environ.get("LITELLM_URL", "http://localhost:4000")
    llm_model = os.environ.get("LLM_MODEL", "default")
    llm_api_key = os.environ.get("LLM_API_KEY", "sk-change-me")

    run_id = str(uuid.uuid4())
    graph = build_pipeline()
    checkpointer = get_checkpointer()
    compiled = graph.compile(checkpointer=checkpointer)

    try:
        result = compiled.invoke(
            {
                "ips_bundle": bundle,
                "run_id": run_id,
                "litellm_url": litellm_url,
                "llm_model": llm_model,
                "llm_api_key": llm_api_key,
            },
            config={"configurable": {"thread_id": run_id}},
        )
    except Exception as e:
        raise HTTPException(status_code=422, detail=f"Care plan generation failed: {e}")

    fhir_bundle = result.get("fhir_bundle", {})
    return Response(
        content=json.dumps(fhir_bundle),
        media_type="application/fhir+json",
        status_code=201,
    )


@app.get("/api/v1/careplans")
def list_careplans(patient: str | None = None, status: str | None = None):
    from acp_writer.nodes.fhir_server_writer import list_care_plans
    return list_care_plans(patient=patient, status=status)


@app.get("/api/v1/careplans/{careplan_id}")
def get_careplan(careplan_id: str):
    from acp_writer.nodes.fhir_server_writer import get_care_plan
    cp = get_care_plan(careplan_id)
    if not cp:
        raise HTTPException(status_code=404, detail=f"Care plan '{careplan_id}' not found")
    return cp


@app.put("/api/v1/careplans/{careplan_id}/status")
async def update_careplan_status(careplan_id: str, request: Request):
    from acp_writer.nodes.fhir_server_writer import approve_care_plan, reject_care_plan
    from acp_writer.services.reviewer import reviewer_from_payload
    data = await request.json()
    new_status = data.get("status")
    if new_status == "active":
        reviewer = reviewer_from_payload(data.get("reviewer"), clinician=data.get("clinician"))
        result = approve_care_plan(careplan_id, reviewer=reviewer)
        if not result:
            raise HTTPException(status_code=404, detail=f"Care plan '{careplan_id}' not found")
        return result
    elif new_status == "entered-in-error":
        reason = data.get("reason", "No reason provided")
        result = reject_care_plan(careplan_id, reason=reason)
        if not result:
            raise HTTPException(status_code=404, detail=f"Care plan '{careplan_id}' not found")
        return result
    else:
        raise HTTPException(status_code=400, detail=f"Invalid status: {new_status}. Use 'active' or 'entered-in-error'.")


# --- Decision Models ---


@app.post("/api/v1/decisions/models", status_code=201)
async def deploy_decision_model(
    request: Request,
    source_cpg: str | None = None,
    validate_only: bool = False,
):
    content_type = request.headers.get("content-type", "")
    body = await request.body()
    dmn_xml = body.decode("utf-8")

    try:
        summary = _parse_dmn_metadata(dmn_xml)
    except ET.ParseError as e:
        raise HTTPException(status_code=400, detail=f"Invalid DMN XML: {e}")

    validation = _validate_dmn_with_engine(dmn_xml)
    if validation is not None and not validation.get("valid", False):
        return _validation_failure_response(validation)
    if validate_only:
        if validation is None:
            raise HTTPException(status_code=503, detail="Decision engine validation unavailable")
        return JSONResponse(status_code=200, content=validation)

    if source_cpg:
        summary.source_cpg = source_cpg

    _dynamic_models[summary.id] = {
        "summary": summary,
        "dmn_xml": dmn_xml,
    }

    logger.info("Deployed decision model: %s (%s, source_cpg=%s)", summary.name, summary.id, source_cpg)
    return summary.model_dump(mode="json")


@app.get("/api/v1/decisions/models")
def list_decision_models():
    return [m["summary"].model_dump(mode="json") for m in _dynamic_models.values()]


@app.get("/api/v1/decisions/models/{model_id}")
def get_decision_model(model_id: str):
    model = _dynamic_models.get(model_id)
    if not model:
        raise HTTPException(status_code=404, detail=f"Model '{model_id}' not found")
    result = model["summary"].model_dump(mode="json")
    result["dmn_xml"] = model["dmn_xml"]
    return result


@app.delete("/api/v1/decisions/models/{model_id}", status_code=204)
def remove_decision_model(model_id: str):
    if model_id not in _dynamic_models:
        raise HTTPException(status_code=404, detail=f"Model '{model_id}' not found")
    del _dynamic_models[model_id]
    logger.info("Removed decision model: %s", model_id)


@app.post("/api/v1/decisions/evaluate/{model_id}")
async def evaluate_decision(model_id: str, request: Request):
    model = _dynamic_models.get(model_id)
    if not model:
        raise HTTPException(status_code=404, detail=f"Model '{model_id}' not found")

    inputs = await request.json()
    try:
        result = _evaluate_jit(model["dmn_xml"], inputs)
        return result
    except Exception as e:
        raise HTTPException(status_code=500, detail=f"Decision evaluation failed: {e}")


# --- Guidelines ---


@app.post("/api/v1/guidelines", status_code=201)
async def register_guideline(request: Request):
    data = await request.json()
    try:
        metadata = CPGMetadata.model_validate(data)
    except Exception as e:
        raise HTTPException(status_code=400, detail=f"Invalid CPG metadata: {e}")
    result = _guidelines_store.register(metadata)
    return result.model_dump(mode="json")


@app.get("/api/v1/guidelines")
def list_guidelines():
    return [g.model_dump(mode="json") for g in _guidelines_store.list_all()]


@app.get("/api/v1/guidelines/{cpg_id}")
def get_guideline(cpg_id: str):
    g = _guidelines_store.get(cpg_id)
    if not g:
        raise HTTPException(status_code=404, detail=f"Guideline '{cpg_id}' not found")
    return g.model_dump(mode="json")


@app.delete("/api/v1/guidelines/{cpg_id}", status_code=204)
def delete_guideline(cpg_id: str):
    if not _guidelines_store.delete(cpg_id):
        raise HTTPException(status_code=404, detail=f"Guideline '{cpg_id}' not found")


# --- Knowledge / Recommendations ---


@app.post("/api/v1/knowledge/recommendations", status_code=201)
async def ingest_recommendation(request: Request):
    data = await request.json()
    try:
        rec = Recommendation.model_validate(data)
    except Exception as e:
        raise HTTPException(status_code=400, detail=f"Invalid recommendation: {e}")
    _vector_store.add(rec)
    return {"id": rec.id, "status": "ingested"}


@app.post("/api/v1/knowledge/recommendations/batch", status_code=201)
async def ingest_recommendation_batch(request: Request):
    data = await request.json()
    try:
        bundle = RecommendationBundle.model_validate(data)
    except Exception as e:
        raise HTTPException(status_code=400, detail=f"Invalid recommendation bundle: {e}")
    _vector_store.add_batch(bundle.recommendations)
    return {
        "source_cpg": bundle.source_cpg,
        "count": len(bundle.recommendations),
        "status": "ingested",
    }


@app.get("/api/v1/knowledge/recommendations")
def list_recommendations(source_cpg: str | None = None):
    recs = _vector_store.list_all(source_cpg=source_cpg)
    return [r.model_dump(mode="json") for r in recs]


@app.get("/api/v1/knowledge/recommendations/{recommendation_id}")
def get_recommendation(recommendation_id: str):
    rec = _vector_store.get(recommendation_id)
    if not rec:
        raise HTTPException(status_code=404, detail=f"Recommendation '{recommendation_id}' not found")
    return rec.model_dump(mode="json")


@app.post("/api/v1/knowledge/search")
async def search_knowledge(request: Request):
    data = await request.json()
    try:
        search_req = RecommendationSearchRequest.model_validate(data)
    except Exception as e:
        raise HTTPException(status_code=400, detail=f"Invalid search request: {e}")
    result = _vector_store.search(search_req)
    return result.model_dump(mode="json")